Fix signed comparisons in utils/vbsp.

+++ sp/src/utils/vbsp/ivp.cpp
@@ -1411,7 +1411,7 @@ static void ConvertModelToPhysCollide( CUtlVector<CPhysCollisionEntry *> &collis
 			}
 		}
 
-		if ( (!numprops || j >= numprops) && numprops < ARRAYSIZE(proplist) )
+		if ( (!numprops || j >= numprops) && numprops < (int) ARRAYSIZE(proplist) )
 		{
 			proplist[numprops].prop = prop;
 			proplist[numprops].area = face->area;
